@@ -6,23 +6,22 @@ plugins {
 
 group = 'org.moddingx'
 
-java.toolchain.languageVersion = JavaLanguageVersion.of(17)
+java.toolchain.languageVersion = JavaLanguageVersion.of(21)
 
 repositories {
     mavenCentral()
-    maven { url = 'https://maven.minecraftforge.net/' }
     maven { url = 'https://maven.moddingx.org/' }
 }
 
 dependencies {
-    implementation 'org.moddingx:LauncherLib:1.1.0'
-    implementation 'org.moddingx:CurseWrapper:3.2'
-    implementation 'org.apache.commons:commons-text:1.10.0'
-    implementation 'com.google.code.gson:gson:2.10.1'
+    implementation 'org.moddingx:LauncherLib:2.1.0'
+    implementation 'org.moddingx:CurseWrapper:5.0'
+    implementation 'org.apache.commons:commons-text:1.12.0'
+    implementation 'com.google.code.gson:gson:2.11.0'
     implementation 'net.sf.jopt-simple:jopt-simple:5.0.4'
-    implementation 'org.jsoup:jsoup:1.16.1'
+    implementation 'org.jsoup:jsoup:1.17.2'
     implementation 'com.moandjiezana.toml:toml4j:0.7.2'
-    implementation 'com.google.guava:guava:31.1-jre'
+    implementation 'com.google.guava:guava:33.2.1-jre'
 }
 
 application.mainClass = 'org.moddingx.updatecheckergenerator.Main'
@@ -76,7 +75,9 @@ publishing {
     }
     repositories {
         maven {
-            url '../moddingx/maven'
+            name 'moddingx'
+            url 'https://maven.moddingx.org/release'
+            credentials(PasswordCredentials)
         }
     }
 }
