@@ -1,10 +1,10 @@
 package org.moddingx.updatecheckergenerator;
 
 import com.google.gson.JsonObject;
-import joptsimple.ArgumentAcceptingOptionSpec;
 import joptsimple.OptionException;
 import joptsimple.OptionParser;
 import joptsimple.OptionSet;
+import joptsimple.OptionSpec;
 import joptsimple.util.PathConverter;
 import joptsimple.util.PathProperties;
 import org.apache.commons.lang3.tuple.Pair;
@@ -24,10 +24,11 @@ public class Main {
 
     public static void main(String[] args) throws IOException {
         OptionParser options = new OptionParser(false);
-        ArgumentAcceptingOptionSpec<Platform> specPlatform = options.acceptsAll(List.of("p", "platform"), "The modding platform to fetch the data from.").withRequiredArg().withValuesConvertedBy(Platform.ARG);
-        ArgumentAcceptingOptionSpec<Path> specCfg = options.acceptsAll(List.of("c", "config"), "A list of project ids to generate update checkers for.").withRequiredArg().withValuesConvertedBy(new PathConverter(PathProperties.FILE_EXISTING));
-        ArgumentAcceptingOptionSpec<Path> specDir = options.acceptsAll(List.of("d", "dir", "directory"), "Output directory.").withRequiredArg().withValuesConvertedBy(new PathConverter());
-        ArgumentAcceptingOptionSpec<Path> specCache = options.acceptsAll(List.of("f", "cache"), "Cache file to use.").withRequiredArg().withValuesConvertedBy(new PathConverter());
+        OptionSpec<Platform> specPlatform = options.acceptsAll(List.of("p", "platform"), "The modding platform to fetch the data from.").withRequiredArg().withValuesConvertedBy(Platform.ARG);
+        OptionSpec<Path> specCfg = options.acceptsAll(List.of("c", "config"), "A list of project ids to generate update checkers for.").withRequiredArg().withValuesConvertedBy(new PathConverter(PathProperties.FILE_EXISTING));
+        OptionSpec<Path> specDir = options.acceptsAll(List.of("d", "dir", "directory"), "Output directory.").withRequiredArg().withValuesConvertedBy(new PathConverter());
+        OptionSpec<Path> specCache = options.acceptsAll(List.of("f", "cache"), "Cache file to use.").withRequiredArg().withValuesConvertedBy(new PathConverter());
+        OptionSpec<Void> specPretty = options.acceptsAll(List.of("pretty"), "Pretty-print the output json.");
         try {
             OptionSet set = options.parse(args);
             if (!set.has(specPlatform) || !set.has(specCfg) || !set.has(specDir)) {
@@ -49,11 +50,13 @@ public class Main {
             FileCache cache = new FileCache();
             if (set.has(specCache)) cache.read(set.valueOf(specCache));
 
+            boolean pretty = set.has(specPretty);
+            
             ModdingPlatform<?> platform = set.valueOf(specPlatform).create();
             projectIds.forEach(projectId -> {
                 try {
                     for (int i = 0; i < 10; i++) {
-                        if (generate(platform, cache, basePath, projectId)) {
+                        if (generate(platform, cache, basePath, projectId, pretty)) {
                             break;
                         }
                         System.out.println("Failed " + (i + 1) + " time(s)");
@@ -71,17 +74,17 @@ public class Main {
         }
     }
 
-    private static boolean generate(ModdingPlatform<?> platform, FileCache cache, Path basePath, String projectId) throws IOException {
+    private static boolean generate(ModdingPlatform<?> platform, FileCache cache, Path basePath, String projectId, boolean pretty) throws IOException {
         try {
             Pair<String, JsonObject> pair = UpdateCheckerGenerator.generateUpdateChecker(platform, projectId, cache);
             Path path = basePath.resolve(pair.getLeft() + ".json");
             Writer writer = Files.newBufferedWriter(path, StandardOpenOption.CREATE, StandardOpenOption.TRUNCATE_EXISTING);
-            writer.write(UpdateCheckerGenerator.GSON.toJson(pair.getRight()) + "\n");
+            writer.write((pretty ? UpdateCheckerGenerator.GSON : UpdateCheckerGenerator.INTERNAL).toJson(pair.getRight()) + "\n");
             writer.close();
             return true;
         } catch (RequestException e) {
             // CurseForge sometimes has problems.
-            // SO we catch the CurseWrapper exception here to let it retry.
+            // So we catch the CurseWrapper exception here to let it retry.
             return false;
         }
     }
